requirements.txt
opencv-python
cvzone
pyttsx3
twilio
cloudinary
python-dotenv
